+++ src/main/java/org/folio/roles/integration/kafka/KafkaMessageListener.java
@@ -6,6 +6,8 @@ import org.folio.roles.integration.kafka.model.ResourceEvent;
 import org.folio.roles.service.capability.CapabilityReplacementsService;
 import org.folio.roles.service.capability.UserPermissionsCacheEvictor;
 import org.folio.spring.context.ExecutionContextBuilder;
+import org.folio.spring.exception.LiquibaseMigrationException;
+import org.folio.spring.liquibase.LiquibaseMigrationLockService;
 import org.folio.spring.scope.FolioExecutionContextSetter;
 import org.folio.spring.service.SystemUserScopedExecutionService;
 import org.springframework.kafka.annotation.KafkaListener;
@@ -21,6 +23,7 @@ public class KafkaMessageListener {
   private final CapabilityReplacementsService capabilityReplacementsService;
   private final SystemUserScopedExecutionService systemUserScopedExecutionService;
   private final UserPermissionsCacheEvictor userPermissionsCacheEvictor;
+  private final LiquibaseMigrationLockService liquibaseMigrationLockService;
 
   /**
    * Handles capability event.
@@ -36,6 +39,8 @@ public class KafkaMessageListener {
     try (
       var ignored = new FolioExecutionContextSetter(executionContextBuilder.buildContext(resourceEvent.getTenant()))) {
       try {
+        checkLiquibaseMigrationRunning(resourceEvent);
+
         systemUserScopedExecutionService.executeSystemUserScoped(() -> {
           var capabilityReplacements = capabilityKafkaEventHandler.handleEvent(resourceEvent);
           capabilityReplacements.ifPresent(capabilityReplacementsService::processReplacements);
@@ -44,6 +49,14 @@ public class KafkaMessageListener {
       } finally {
         userPermissionsCacheEvictor.evictUserPermissionsForCurrentTenant();
       }
+    }
+  }
+
+  private void checkLiquibaseMigrationRunning(ResourceEvent resourceEvent) {
+    if (liquibaseMigrationLockService.isMigrationRunning()) {
+      log.warn("Liquibase migration in progress for tenant: {}", resourceEvent.getTenant());
+      throw new LiquibaseMigrationException(
+        "Liquibase migration is still running for tenant: " + resourceEvent.getTenant());
     }
   }
 }